Исправить замечения.
