+++ gateway/src/main/java/ru/practicum/shareit/exception/ErrorHandler.java
@@ -1,6 +1,8 @@
 package ru.practicum.shareit.exception;
 
 import lombok.extern.slf4j.Slf4j;
+import org.springframework.core.Ordered;
+import org.springframework.core.annotation.Order;
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.ControllerAdvice;
 import org.springframework.web.bind.annotation.ExceptionHandler;
@@ -14,6 +16,30 @@ public class ErrorHandler {
     @ExceptionHandler
     @ResponseStatus(HttpStatus.BAD_REQUEST)
     public ErrorResponse handle(NoSuchStateForBookingSearchException e) {
+        log.error(e.getMessage());
+        return new ErrorResponse(e.getMessage());
+    }
+
+    @ExceptionHandler
+    @Order(Ordered.LOWEST_PRECEDENCE)
+    @ResponseStatus(HttpStatus.INTERNAL_SERVER_ERROR)
+    public ErrorResponse handle(RuntimeException e) {
+        String message = "Произошла ошибка: " + e.getMessage() + " " + e;
+        log.error(message);
+        return new ErrorResponse(message);
+    }
+
+    @ExceptionHandler
+    @ResponseStatus(HttpStatus.INTERNAL_SERVER_ERROR)
+    public ErrorResponse handle(Error e) {
+        String message = "Произошла ошибка: " + e.getMessage() + " " + e;
+        log.error(message);
+        return new ErrorResponse(message);
+    }
+
+    @ExceptionHandler
+    @ResponseStatus(HttpStatus.BAD_REQUEST)
+    public ErrorResponse handle(InvalidDateException e) {
         log.error(e.getMessage());
         return new ErrorResponse(e.getMessage());
     }
