+++ server/src/main/java/ru/practicum/shareit/booking/BookingController.java
@@ -2,16 +2,12 @@ package ru.practicum.shareit.booking;
 
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.*;
 
-import javax.validation.constraints.Positive;
-import javax.validation.constraints.PositiveOrZero;
 import java.util.List;
 
 @RequiredArgsConstructor
 @RestController
-@Validated
 @RequestMapping(path = "/bookings")
 @Slf4j
 public class BookingController {
@@ -27,8 +23,8 @@ public class BookingController {
     @GetMapping
     public List<SavedBookingDto> getBookingsOfUser(@RequestParam(defaultValue = "ALL") String state,
                                                    @RequestHeader("X-Sharer-User-Id") long userId,
-                                                   @RequestParam(defaultValue = "0") @PositiveOrZero long from,
-                                                   @RequestParam(defaultValue = "20") @Positive long size) {
+                                                   @RequestParam(defaultValue = "0") long from,
+                                                   @RequestParam(defaultValue = "20") long size) {
         log.info("получен запрос на получение бронирований со статусом " + state + " пользователя с id " + userId);
         return bookingService.getBookingsOfUser(from, size, state, userId);
     }
@@ -36,8 +32,8 @@ public class BookingController {
     @GetMapping("owner")
     public List<SavedBookingDto> getBookingsOfItemsOwner(@RequestParam(defaultValue = "ALL") String state,
                                                          @RequestHeader("X-Sharer-User-Id") long userId,
-                                                         @RequestParam(defaultValue = "0") @PositiveOrZero long from,
-                                                         @RequestParam(defaultValue = "20") @Positive long size) {
+                                                         @RequestParam(defaultValue = "0") long from,
+                                                         @RequestParam(defaultValue = "20") long size) {
         log.info("получен запрос на получение бронирований от хозяина вещей со статусом " + state +
                 " пользователя с id " + userId);
         return bookingService.getBookingsOfItemsOwner(from, size, state, userId);
