+++ server/src/main/java/ru/practicum/shareit/exception/ErrorHandler.java
@@ -53,13 +53,6 @@ public class ErrorHandler {
     }
 
     @ExceptionHandler
-    @ResponseStatus(HttpStatus.BAD_REQUEST)
-    public ErrorResponse handle(InvalidDateException e) {
-        log.error(e.getMessage());
-        return new ErrorResponse(e.getMessage());
-    }
-
-    @ExceptionHandler
     @ResponseStatus(HttpStatus.NOT_FOUND)
     public ErrorResponse handle(BookingNotBelongException e) {
         log.error(e.getMessage());
