+++ server/src/main/java/ru/practicum/shareit/item/ItemDto.java
@@ -6,9 +6,6 @@ import lombok.Data;
 import lombok.NoArgsConstructor;
 import ru.practicum.shareit.booking.BookingItemDto;
 
-import javax.validation.constraints.NotBlank;
-import javax.validation.constraints.NotEmpty;
-import javax.validation.constraints.NotNull;
 import java.util.List;
 
 @AllArgsConstructor
@@ -17,16 +14,8 @@ import java.util.List;
 @NoArgsConstructor
 public class ItemDto {
     private long id;
-
-    @NotEmpty
-    @NotBlank
     private String name;
-
-    @NotEmpty
-    @NotBlank
     private String description;
-
-    @NotNull
     private Boolean available;
     private BookingItemDto lastBooking;
     private BookingItemDto nextBooking;
