+++ server/src/main/java/ru/practicum/shareit/request/ItemRequestController.java
@@ -2,17 +2,13 @@ package ru.practicum.shareit.request;
 
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.*;
 
-import javax.validation.constraints.Positive;
-import javax.validation.constraints.PositiveOrZero;
 import java.util.List;
 
 @RequiredArgsConstructor
 @RestController
 @Slf4j
-@Validated
 @RequestMapping(path = "/requests")
 public class ItemRequestController {
     private final ItemRequestService itemRequestService;
@@ -39,8 +35,8 @@ public class ItemRequestController {
 
     @GetMapping("all")
     public List<ItemRequestDto> getAllRequests(@RequestHeader("X-Sharer-User-Id") long userId,
-                                               @RequestParam(defaultValue = "0") @PositiveOrZero long from,
-                                               @RequestParam(defaultValue = "20") @Positive long size) {
+                                               @RequestParam(defaultValue = "0") long from,
+                                               @RequestParam(defaultValue = "20") long size) {
         log.info("получен запрос на получение всех запросов от пользователя с id " + userId);
         return itemRequestService.getAllRequests(from, size, userId);
     }
