+++ server/src/main/java/ru/practicum/shareit/user/UserController.java
@@ -2,7 +2,6 @@ package ru.practicum.shareit.user;
 
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
@@ -11,7 +10,6 @@ import java.util.List;
 @RestController
 @RequestMapping(path = "/users")
 @Slf4j
-@Validated
 public class UserController {
     private final UserService userService;
